push

+++ Routes/UserContact.js
@@ -27,7 +27,7 @@ ContactRouter.get("/:n",adminAuth, async(req,res)=>{
     }
 })
 
-ContactRouter.get("/searchbyemail/:email",adminAuth, async(req,res)=>{
+ContactRouter.get("/searchbyemail/:email",adminAuth,async(req,res)=>{
   const {email}=req.params;
 try {
     let data= await ContactModel.find({email});
@@ -64,21 +64,39 @@ ContactRouter.get("/searchbyname/:name",adminAuth, async(req,res)=>{
 ContactRouter.post("/add",adminAuth, async(req,res)=>{
      const {name,number,email}=req.body;
 
-     const d= await ContactModel.find({email});
-     console.log(d);
-     if(d.length>0){
-        res.status(400).send({"msg":"Email Id Alredy Exist try another"})
+     if(req.body){
+        const d= await ContactModel.find({email});
+        console.log(d);
+        if(d.length>0){
+           res.status(400).send({"msg":"Email Id Alredy Exist try another"})
+        }else{
+           let data= ContactModel(req.body);  
+           await data.save()     
+           res.status(200).send({"msg":"SuccesFully Added"})
+        }
+     }else{
+      res.send("Error")
      }
+
      
-     let data= ContactModel(req.body);  
-     await data.save()     
-     res.status(200).send({"msg":"SuccesFully Added"})
+     
+    
 })         
 
 ContactRouter.delete("/delete/:id",adminAuth, async(req,res)=>{
   let {id}=req.params;
-  let data= await ContactModel.findByIdAndDelete({_id:id})
-  res.send({"msg":"SucessFully Deleted"})
+  if(id){
+    try {
+    let data= await ContactModel.findByIdAndDelete({_id:id})
+    res.send({"msg":"SucessFully Deleted"})
+    } catch (error){
+        res.status(400).send({"msg":"id not found"})
+    }
+    
+  }else{
+    res.send({"msg":"No Id Present"})
+  }
+  
   
   
 })
@@ -86,8 +104,18 @@ ContactRouter.delete("/delete/:id",adminAuth, async(req,res)=>{
 
 ContactRouter.patch("/edit/:id",adminAuth, async(req,res)=>{
     let {id}=req.params;
-    let data= await ContactModel.findByIdAndUpdate({_id:id},req.body);
-    res.send({"Msg":"Sucessfully edited"});
+    if(id){
+        try {
+            let data= await ContactModel.findByIdAndUpdate({_id:id},req.body);
+        res.send({"Msg":"Sucessfully edited"});
+        } catch (error) {
+            res.status(400).send({"msg":"id not found"})
+        }
+        
+    }else{
+        res.send({"msg":"No ID Present"})
+    }
+   
     
 })
